requirements.txt
flask
flask_sqlalchemy
psycopg2
jsonify
gunicorn
python-dotenv